% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{gen_news}
\alias{gen_news}
\title{Generate the news between two data releases}
\usage{
gen_news(model, target_period, old_data, new_data)
}
\arguments{
\item{model}{a trained LSTM model gotten from calling LSTM()}

\item{target_period}{string in "YYYY-MM-DD", target prediction date}

\item{old_data}{dataframe of previous dataset}

\item{new_data}{dataframe of new dataset}
}
\value{

}
\description{
Generate the news between two data releases using the method of holding out new data feature by feature and recording the differences in model output
}
