% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{initialize_session}
\alias{initialize_session}
\title{Initialize Python libraries}
\usage{
initialize_session(python_path = "")
}
\arguments{
\item{python_path}{string, location of Python. If left empty, uses default.}
}
\description{
Initializes Python libraries for the session
}
