% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{install_python_libraries}
\alias{install_python_libraries}
\title{Install required Python libraries}
\usage{
install_python_libraries()
}
\description{
Installs required Python libraries. If this fails, install `pandas`, `numpy`, `pmdarima`, `dill`, `torch` and `nowcast-lstm` manually from the command line with pip
}
