% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{load_lstm}
\alias{load_lstm}
\title{Load a trained LSTM model from disk}
\usage{
load_lstm(path, python_model_name = "load_tmp_model")
}
\arguments{
\item{path}{the file name and path to read the model from. Should end in ".pkl"}

\item{python_model_name}{what the model will be called in the python session. Relevant if more than one model is being loaded. For clarity, should be the same as the name of the R object the model is being saved to.}
}
\value{
trained LSTM model
}
\description{
Load a trained LSTM model from disk
}
