% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{predict}
\alias{predict}
\title{Get predictions from a trained LSTM model}
\usage{
predict(model, data, only_actuals_obs = FALSE)
}
\arguments{
\item{model}{a trained LSTM model gotten from calling LSTM()}

\item{data}{dataframe with the same columns the model was trained on}

\item{only_actuals_obs}{whether or not to produce predictions for periods without an actual. E.g. FALSE will return predictions for months in between quarters, even if the target variable is quarterly.}
}
\value{
dataframe with periods, actuals if available, and predictions
}
\description{
Get predictions from a trained LSTM model
}
