% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{ragged_preds}
\alias{ragged_preds}
\title{Get predictions on artificial vintages}
\usage{
ragged_preds(model, pub_lags, lag, data, start_date = NULL, end_date = NULL)
}
\arguments{
\item{model}{a trained LSTM model gotten from calling LSTM()}

\item{pub_lags}{list of integers, list of periods back each input variable is set to missing. I.e. publication lag of the variable.}

\item{lag}{integer, simulated periods back. E.g. -2 = simulating data as it would have been 2 months before target period, 1 = 1 month after, etc.}

\item{data}{dataframe to generate the ragged datasets and predictions on}

\item{start_date}{string in "YYYY-MM-DD" format, start date of generating ragged preds. To save calculation time, i.e. just calculating after testing date instead of all dates}

\item{end_date}{string in "YYYY-MM-DD" format, end date of generating ragged preds}
}
\value{
dataframe of actuals, predictions, and dates
}
\description{
Get predictions on artificial vintages
}
