% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_wrapper.R
\name{save_lstm}
\alias{save_lstm}
\title{Save a trained LSTM model to disk}
\usage{
save_lstm(model, path, python_model_name = "")
}
\arguments{
\item{model}{a trained LSTM model gotten from calling LSTM()}

\item{path}{the file name and path to save the model to. Should end in ".pkl"}

\item{python_model_name}{what the model is called in the python session. Relevant if more than one model is in use. Defaults to same name used for the R object. For clarity, should be the same as the name of the R object the model is being saved to when the `load_lstm` and `LSTM` functions are initially used.}
}
\value{
trained LSTM model
}
\description{
Save a trained LSTM model to disk
}
